Hemera_OS/pentaia/zios/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-genai
vecs
pydantic-settings
python-multipart
psycopg2-binary
sqlalchemy